added testing from modules altered setup.py tags
