+++ setup.py
@@ -34,19 +34,17 @@ setup(
     install_requires=requirements,
     license="MIT license",
     zip_safe=False,
-    keywords='hydrofunctions',
+    keywords='hydrofunctions hydrology USGS stream gauge water NWIS',
     classifiers=[
-        'Development Status :: 2 - Pre-Alpha',
-        'Intended Audience :: Developers',
+        'Development Status :: 1 - Planning',
+        'Intended Audience :: Science/Research',
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
-        "Programming Language :: Python :: 2",
-        'Programming Language :: Python :: 2.6',
-        'Programming Language :: Python :: 2.7',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.3',
         'Programming Language :: Python :: 3.4',
         'Programming Language :: Python :: 3.5',
+        'Topic :: Scientific/Engineering :: Information Analysis',
+        'Topic :: Scientific/Engineering :: Visualization',
+        'Topic :: Utilities'
     ],
     test_suite='tests',
     tests_require=test_requirements
