+++ tests/test_hydrofunctions.py
@@ -34,22 +34,7 @@ class TestHydrofunctions(unittest.TestCase):
         self.assertEqual(expected, actual, msg="first() did not return True.")
 
 
-def run_cmd():
-    import subprocess
-    ## command to run  ##
-    cmd = "nosetests"
 
-    ## run it ##
-    p = subprocess.Popen(cmd, shell=True, stderr=subprocess.PIPE)
-
-    ## But do not wait till netstat finish, start displaying output immediately ##
-    while True:
-        out = p.stderr.read(1)
-        if out == '' and p.poll() != None:
-            break
-        if out != '':
-            sys.stdout.write(out)
-            sys.stdout.flush()
 
 
 if __name__ == '__main__':
